Add function that gets a voter's rewards for all groups (commit before rebase)
