+++ apps/explorer/test/explorer/celo/voter_rewards_for_group_test.exs
@@ -1,0 +1,82 @@
+defmodule Explorer.Celo.VoterRewardsForGroupTest do
+  use Explorer.DataCase
+
+  alias Explorer.Celo.VoterRewardsForGroup
+  alias Explorer.Chain.Hash
+  alias Explorer.SetupVoterRewardsTest
+
+  describe "calculate/2" do
+    test "returns all rewards for a voter voting for a specific group" do
+      {voter_address_1_hash, group_address_hash} = SetupVoterRewardsTest.setup()
+
+      {:ok, rewards} =
+        VoterRewardsForGroup.calculate(Hash.to_string(voter_address_1_hash), Hash.to_string(group_address_hash))
+
+      assert rewards ==
+               %{
+                 total: 350,
+                 epochs: [
+                   %{amount: 80, date: ~U[2022-01-01 17:42:43.162804Z], epoch_number: 619},
+                   %{amount: 20, date: ~U[2022-01-02 17:42:43.162804Z], epoch_number: 620},
+                   %{amount: 75, date: ~U[2022-01-03 17:42:43.162804Z], epoch_number: 621},
+                   %{amount: 31, date: ~U[2022-01-04 17:42:43.162804Z], epoch_number: 622},
+                   %{amount: 77, date: ~U[2022-01-05 17:42:43.162804Z], epoch_number: 623},
+                   %{amount: 67, date: ~U[2022-01-06 17:42:43.162804Z], epoch_number: 624}
+                 ]
+               }
+    end
+  end
+
+  describe "amount_activated_or_revoked_last_day/2" do
+    test "sums a voter's activated and revoked CELO for the previous day of the block passed" do
+      voter_activated_or_revoked = [
+        %{
+          block_hash: %Hash{
+            byte_count: 32,
+            bytes: <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>
+          },
+          block_number: 10_692_863,
+          amount_activated_or_revoked: 650,
+          event: "0x45aac85f38083b18efe2d441a65b9c1ae177c78307cb5a5d4aec8f7dbcaeabfe",
+          group_hash: "0x0000000000000000000000000000000000000003",
+          voter_hash: "0x0000000000000000000000000000000000000001"
+        },
+        %{
+          block_hash: %Hash{
+            byte_count: 32,
+            bytes: <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8>>
+          },
+          block_number: 10_695_496,
+          amount_activated_or_revoked: 650,
+          event: "0x45aac85f38083b18efe2d441a65b9c1ae177c78307cb5a5d4aec8f7dbcaeabfe",
+          group_hash: "0x0000000000000000000000000000000000000003",
+          voter_hash: "0x0000000000000000000000000000000000000001"
+        },
+        %{
+          block_hash: %Hash{
+            byte_count: 32,
+            bytes: <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16>>
+          },
+          block_number: 10_706_524,
+          amount_activated_or_revoked: 350,
+          event: "0xae7458f8697a680da6be36406ea0b8f40164915ac9cc40c0dad05a2ff6e8c6a8",
+          group_hash: "0x0000000000000000000000000000000000000003",
+          voter_hash: "0x0000000000000000000000000000000000000001"
+        },
+        %{
+          block_hash: %Hash{
+            byte_count: 32,
+            bytes: <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16>>
+          },
+          block_number: 10_796_524,
+          amount_activated_or_revoked: 350,
+          event: "0xae7458f8697a680da6be36406ea0b8f40164915ac9cc40c0dad05a2ff6e8c6a8",
+          group_hash: "0x0000000000000000000000000000000000000003",
+          voter_hash: "0x0000000000000000000000000000000000000001"
+        }
+      ]
+
+      assert VoterRewardsForGroup.amount_activated_or_revoked_last_day(voter_activated_or_revoked, 10_710_000) == 950
+    end
+  end
+end
