+++ apps/explorer/test/explorer/chain_test.exs
@@ -6088,78 +6088,108 @@ defmodule Explorer.ChainTest do
     end
   end
 
-  describe "voter_rewards_for_group/2" do
-    test "returns all rewards for a voter voting for a specific group" do
-      {voter_address_1_hash, group_address_hash} = SetupVoterRewardsTest.setup()
+  describe "voter_rewards/1" do
+    test "returns all rewards for a voter" do
+      [validator_group_active_vote_revoked, _, validator_group_vote_activated, _] = Events.voter_events()
+      %Address{hash: voter_address_1_hash} = voter_address_1 = insert(:address)
+      %Address{hash: voter_address_2_hash} = voter_address_2 = insert(:address)
+      %Address{hash: group_address_1_hash} = group_address_1 = insert(:address)
+      %Address{hash: group_address_2_hash} = group_address_2 = insert(:address)
+
+
+      block_1 = insert(:block, number: 10_692_863, timestamp: ~U[2022-01-01 13:08:43.162804Z])
+
+      transaction_1 =
+        :transaction
+        |> insert(from_address: voter_address_1)
+        |> with_block(block_1)
+
+      insert(:log,
+        address: voter_address_1,
+        block: transaction_1.block,
+        block_number: transaction_1.block_number,
+        data: Chain.raw_abi_encode_integers([650, 10000]),
+        first_topic: validator_group_vote_activated,
+        second_topic: to_string(voter_address_1_hash),
+        third_topic: to_string(group_address_1_hash),
+        index: 1,
+        transaction: transaction_1
+      )
+
+      block_2 = insert(:block, number: 10744703, timestamp: ~U[2022-01-04 13:08:43.162804Z])
+
+      transaction_2 =
+        :transaction
+        |> insert(from_address: voter_address_1)
+        |> with_block(block_2)
+
+      insert(:log,
+        address: voter_address_1,
+        block: transaction_2.block,
+        block_number: transaction_2.block_number,
+        data: Chain.raw_abi_encode_integers([650, 10000]),
+        first_topic: validator_group_vote_activated,
+        second_topic: to_string(voter_address_1_hash),
+        third_topic: to_string(group_address_2_hash),
+        index: 1,
+        transaction: transaction_2
+      )
+
+      block_3 = insert(:block, number: 10779263, timestamp: ~U[2022-01-06 13:08:43.162804Z])
+
+      transaction_3 =
+        :transaction
+        |> insert(from_address: voter_address_1)
+        |> with_block(block_3)
+
+      insert(:log,
+        address: voter_address_2,
+        block: transaction_3.block,
+        block_number: transaction_3.block_number,
+        data: Chain.raw_abi_encode_integers([650, 10000]),
+        first_topic: validator_group_vote_activated,
+        second_topic: to_string(voter_address_2_hash),
+        third_topic: to_string(group_address_2_hash),
+        index: 1,
+        transaction: transaction_3
+      )
 
       {:ok, rewards} =
-        Chain.voter_rewards_for_group(Hash.to_string(voter_address_1_hash), Hash.to_string(group_address_hash))
+        Chain.voter_rewards(Hash.to_string(voter_address_1_hash))
 
       assert rewards ==
                %{
-                 total: 350,
-                 epochs: [
-                   %{amount: 80, date: ~U[2022-01-01 17:42:43.162804Z], epoch_number: 619},
-                   %{amount: 20, date: ~U[2022-01-02 17:42:43.162804Z], epoch_number: 620},
-                   %{amount: 75, date: ~U[2022-01-03 17:42:43.162804Z], epoch_number: 621},
-                   %{amount: 31, date: ~U[2022-01-04 17:42:43.162804Z], epoch_number: 622},
-                   %{amount: 77, date: ~U[2022-01-05 17:42:43.162804Z], epoch_number: 623},
-                   %{amount: 67, date: ~U[2022-01-06 17:42:43.162804Z], epoch_number: 624}
+                 total_reward_celo: 350,
+                 voter_account: to_string(voter_address_1_hash),
+                 from: ~U[2022-01-03 00:00:00.000000Z],
+                 to: ~U[2022-01-06 00:00:00.000000Z],
+                 rewards: [
+                   %{
+                     amount: 80,
+                     date: ~U[2022-01-01 17:42:43.162804Z],
+                     block_number: 619,
+                     block_hash: 619,
+                     epoch_number: 619,
+                     group: to_string(group_address_1_hash)
+                   },
+                   %{
+                     amount: 20,
+                     date: ~U[2022-01-02 17:42:43.162804Z],
+                     block_number: 620,
+                     block_hash: 620,
+                     epoch_number: 620,
+                     group: to_string(group_address_2_hash)
+                   },
+                   %{
+                     amount: 75,
+                     date: ~U[2022-01-03 17:42:43.162804Z],
+                     block_number: 621,
+                     block_hash: 621,
+                     epoch_number: 621,
+                     group: to_string(group_address_2_hash)
+                   }
                  ]
                }
-    end
-  end
-
-  describe "amount_activated_or_revoked_last_day/2" do
-    test "sums a voter's activated and revoked CELO for the previous day of the block passed" do
-      voter_activated_or_revoked = [
-        %{
-          block_hash: %Explorer.Chain.Hash{
-            byte_count: 32,
-            bytes: <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>
-          },
-          block_number: 10_692_863,
-          amount_activated_or_revoked: 650,
-          event: "0x45aac85f38083b18efe2d441a65b9c1ae177c78307cb5a5d4aec8f7dbcaeabfe",
-          group_hash: "0x0000000000000000000000000000000000000003",
-          voter_hash: "0x0000000000000000000000000000000000000001"
-        },
-        %{
-          block_hash: %Explorer.Chain.Hash{
-            byte_count: 32,
-            bytes: <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 8>>
-          },
-          block_number: 10_695_496,
-          amount_activated_or_revoked: 650,
-          event: "0x45aac85f38083b18efe2d441a65b9c1ae177c78307cb5a5d4aec8f7dbcaeabfe",
-          group_hash: "0x0000000000000000000000000000000000000003",
-          voter_hash: "0x0000000000000000000000000000000000000001"
-        },
-        %{
-          block_hash: %Explorer.Chain.Hash{
-            byte_count: 32,
-            bytes: <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16>>
-          },
-          block_number: 10_706_524,
-          amount_activated_or_revoked: 350,
-          event: "0xae7458f8697a680da6be36406ea0b8f40164915ac9cc40c0dad05a2ff6e8c6a8",
-          group_hash: "0x0000000000000000000000000000000000000003",
-          voter_hash: "0x0000000000000000000000000000000000000001"
-        },
-        %{
-          block_hash: %Explorer.Chain.Hash{
-            byte_count: 32,
-            bytes: <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 16>>
-          },
-          block_number: 10_796_524,
-          amount_activated_or_revoked: 350,
-          event: "0xae7458f8697a680da6be36406ea0b8f40164915ac9cc40c0dad05a2ff6e8c6a8",
-          group_hash: "0x0000000000000000000000000000000000000003",
-          voter_hash: "0x0000000000000000000000000000000000000001"
-        }
-      ]
-
-      assert Chain.amount_activated_or_revoked_last_day(voter_activated_or_revoked, 10_710_000) == 950
     end
   end
 end
